requirements.txt
requests==2.18.0  # Deze versie heeft bekende kwetsbaarheden
python-dotenv
